@@ -1,9 +1,10 @@
 package com.davidmiguel.sample;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -17,7 +18,7 @@ public class Card3Fragment extends Fragment {
         // Required empty constructor
     }
     
-    public static Card3Fragment newInstance() {
+    static Card3Fragment newInstance() {
         return new Card3Fragment();
     }
 
