@@ -1,10 +1,11 @@
 package com.davidmiguel.sample;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
 import com.davidmiguel.dragtoclose.DragToClose;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * In this example the card can be dragged out card but the activity is not closed.
